♻️ refactor: 모달 종류들 /common/modal-contents 로 폴더 이동 모달종류가 늘어나서 옮겼습니다.
