+++ src/components/common/modal-contents/OnlyTextContent.tsx
