+++ src/components/common/modal-contents/WarningContent.tsx
